Code cleanup, 100 lines less

+++ iPA Edit.py
@@ -2,20 +2,73 @@ import os
 import zipfile 
 import plistlib 
 import shutil
+import pickle
+import subprocess
+import time
  
 payload_name = "Payload.zip" 
+filename = 'variable.pkl'
 
+
+def clear_terminal():
+    if os.name == 'nt': # Windows
+        os.system('cls')
+    elif os.name == 'posix': # MacOS oder Linux
+        os.system('clear')
+        
+                                                                                #unzip iPA
+def unzip_ipa(ipa_path):
+    clear_terminal()
+    file_name_no_ipa = os.path.basename(ipa_path)[:-4]
+    zip_path = ipa_path.replace(".ipa", ".zip")
+
+    if os.path.exists(ipa_path):
+        os.rename(ipa_path, zip_path)
+        print("iPA file successfully renamed to .Zip")
+        time.sleep(3)
+    else:
+        print("The .iPA file could not be found. Try again...")
+        exit()
+
+    with zipfile.ZipFile(zip_path, 'r') as zip_ref:
+        zip_ref.extractall(os.path.dirname(zip_path))
+
+    payload_path = os.path.join(os.path.dirname(zip_path), "Payload")
+    print(payload_path)
+    app_folder = os.listdir(payload_path)[0]
+    app_path = os.path.join(payload_path, app_folder)
+    return app_path, file_name_no_ipa, zip_path
+                                                                                #zip iPA
+def zip_ipa(ipa_path, app_path, file_name_no_ipa):
+    zip_path = ipa_path.replace(".ipa", ".zip")
+    if os.path.exists(zip_path):
+        os.remove(zip_path)
+
+    with zipfile.ZipFile(zip_path, 'w', zipfile.ZIP_DEFLATED) as zip_file:
+        for root, dirs, files in os.walk(app_path):
+            for file in files:
+                file_path = os.path.join(root, file)
+                zip_file.write(file_path, file_path.replace(app_path, ""))
+
+    os.rename(zip_path, file_name_no_ipa + ".ipa")
+
+                                                                                            #start
+                                                                                            
+print("[1] change Bundle-ID") 
+print("[2] change App-Name") 
+print("[3] change App-Version") 
+print("[4] change App-Icon")
+print("[5] inject Satella Jailed")
+print("[6] inject .debs/.dylibs")
  
-def bundleID(): 
-     
+option = int(input("choose an option: \n"))
+clear_terminal() 
  
-    with zipfile.ZipFile(zip_path, 'r') as zip_ref: 
-        zip_ref.extractall(os.path.dirname(zip_path)) 
- 
-    payload_path = os.path.join(os.path.dirname(zip_path), "Payload") 
-    print(payload_path) 
-    app_folder = os.listdir(payload_path)[0] 
-    app_path = os.path.join(payload_path, app_folder) 
+if option == 1: 
+    
+    ipa_path = input("Please enter the path to the IPA file:\n ")
+    clear_terminal()
+    app_path, file_name_no_ipa, zip_path = unzip_ipa(ipa_path)
  
     info_plist_path = os.path.join(app_path, "Info.plist") 
     with open(info_plist_path, 'rb') as fp: 
@@ -24,46 +77,24 @@ def bundleID():
     old_bundle_id = pl['CFBundleIdentifier'] 
     print("current Bundle-ID: ", old_bundle_id) 
  
-    new_bundle_id = input("please enter your new Bundle-ID: ") 
-    pl['CFBundleIdentifier'] = new_bundle_id 
+    new_bundle_id = input("\nplease enter your new Bundle-ID: \n") 
+    clear_terminal()
+    pl['CFBundleIdentifier'] = new_bundle_id
+    print("Patching.... Pls wait")
  
     with open(info_plist_path, 'wb') as fp: 
          plistlib.dump(pl, fp) 
     
-    payload_path2 = payload_path
+    zip_ipa(ipa_path, app_path, file_name_no_ipa)
+    clear_terminal()
+    print("The Bundle ID was changed successfully.")
+ 
+if option == 2: 
     
-    if os.path.basename(payload_path2) == "Payload":
-        output_path = payload_path2[:len(payload_path2)-len("/Payload")]
-    else:
-        output_path = os.path.dirname(payload_path2)
- 
-    output_path = os.path.join(output_path) 
-    with zipfile.ZipFile(os.path.join(output_path, "Payload.zip"), 'w', zipfile.ZIP_DEFLATED) as zip_file:  
-       for root, dirs, files in os.walk(payload_path): 
-           for file in files: 
-               file_path = os.path.join(root, file) 
-               zip_file.write(file_path, file_path.replace(payload_path, "Payload")) 
-                
-                
-    user_new_ipa_name = input(f"\n\nenter a new name for your edited .ipa (without the .ipa at the end) \noriginal .ipa name: {file_name_no_ipa}'\n") 
-    user_new_ipa_name = user_new_ipa_name.strip() + ".ipa"
-    os.rename(payload_path+".zip", user_new_ipa_name) 
-    edited_file_path = os.path.join(output_path, user_new_ipa_name)
-    os.replace(user_new_ipa_name, edited_file_path)
-
- 
-    print("The bundle ID was changed successfully.") 
- 
-def changeName(): 
- 
-    with zipfile.ZipFile(zip_path, 'r') as zip_ref: 
-        zip_ref.extractall(os.path.dirname(zip_path)) 
- 
-    payload_path = os.path.join(os.path.dirname(zip_path), "Payload") 
-    print(payload_path) 
-    app_folder = os.listdir(payload_path)[0] 
-    app_path = os.path.join(payload_path, app_folder) 
- 
+    ipa_path = input("Please enter the path to the IPA file:\n ")
+    clear_terminal()
+    app_path, file_name_no_ipa, zip_path = unzip_ipa(ipa_path)
+    
     info_plist_path = os.path.join(app_path, "Info.plist") 
     with open(info_plist_path, 'rb') as fp: 
             pl = plistlib.load(fp) 
@@ -71,58 +102,64 @@ def changeName():
     old_display_name = pl['CFBundleDisplayName'] 
     print("current App-Name: ", old_display_name) 
  
-    new_display_name = input("please enter your new App-Name: ") 
+    new_display_name = input("please enter your new App-Name: \n")
+    clear_terminal() 
     pl['CFBundleDisplayName'] = new_display_name 
+    print("Patching.... Pls wait")
  
     with open(info_plist_path, 'wb') as fp: 
          plistlib.dump(pl, fp) 
- 
          
-    payload_path2 = payload_path
+    zip_ipa(ipa_path, app_path, file_name_no_ipa)
+    clear_terminal()
+    print("The App Name was changed successfully.")
     
-    if os.path.basename(payload_path2) == "Payload":
-        output_path = payload_path2[:len(payload_path2)-len("/Payload")]
-    else:
-        output_path = os.path.dirname(payload_path2) 
+if option == 3:
+    
+    ipa_path = input("Please enter the path to the IPA file:\n ")
+    clear_terminal()
+    app_path, file_name_no_ipa, zip_path = unzip_ipa(ipa_path)
+
+    info_plist_path = os.path.join(app_path, "Info.plist") 
+    with open(info_plist_path, 'rb') as fp: 
+            pl = plistlib.load(fp) 
  
-    output_path = os.path.join(output_path) 
-    with zipfile.ZipFile(os.path.join(output_path, "Payload.zip"), 'w', zipfile.ZIP_DEFLATED) as zip_file:  
-       for root, dirs, files in os.walk(payload_path): 
-           for file in files: 
-               file_path = os.path.join(root, file) 
-               zip_file.write(file_path, file_path.replace(payload_path, "Payload")) 
-                
-                
-    user_new_ipa_name = input(f"\n\nenter a new name for your edited .ipa (without the .ipa at the end) \noriginal .ipa name: {file_name_no_ipa}'\n") 
-    user_new_ipa_name = user_new_ipa_name.strip() + ".ipa"
-    os.rename(payload_path+".zip", user_new_ipa_name) 
-    edited_file_path = os.path.join(output_path, user_new_ipa_name)
-    os.replace(user_new_ipa_name, edited_file_path)
+    old_version = pl['CFBundleShortVersionString'] 
+    print("\ncurrent App-Version: ", old_version) 
  
-    print("The App-Name was changed successfully.") 
+    new_version = input("\nplease enter your new App-Version: \n")
+    clear_terminal()
+    pl['CFBundleShortVersionString'] = new_version
+    print("Patching.... Pls wait") 
  
-def changeIcon(): 
-     
-    with zipfile.ZipFile(zip_path, 'r') as zip_ref: 
-        zip_ref.extractall(os.path.dirname(zip_path)) 
- 
-    payload_path = os.path.join(os.path.dirname(zip_path), "Payload") 
-    print(payload_path) 
-    app_folder = os.listdir(payload_path)[0] 
-    app_path = os.path.join(payload_path, app_folder) 
- 
+    with open(info_plist_path, 'wb') as fp: 
+         plistlib.dump(pl, fp) 
+         
+    zip_ipa(ipa_path, app_path, file_name_no_ipa)
+
+    print("The App Version was changed successfully.")
+    
+if option == 4:
+    
+    ipa_path = input("Please enter the path to the IPA file:\n ")
+    clear_terminal()
+    app_path, file_name_no_ipa, zip_path = unzip_ipa(ipa_path)
+    
     directory = app_path # aktuellen ordner 
     for filename in os.listdir(directory): 
         if filename.startswith("AppIcon") and filename.endswith(".png"): 
             file_path = os.path.join(directory, filename) 
             try: 
-                os.remove(file_path) 
-                print(f"icon '{filename}' deleted successfully.") 
+                os.remove(file_path)
+                clear_terminal()
+                print(f"\nIcon '{filename}' deleted successfully.") 
+                clear_terminal()
             except Exception as e: 
-                print(f"Error deleting file '{filename}': {e}") 
+                print(f"\nError deleting file '{filename}': {e}") 
                  
-    icon_path = input("\n\nEnter the path to the .png you want to use as App-Icon:\n")
-    
+    icon_path = input("Enter the path to the .png you want to use as App-Icon:\n")
+    clear_terminal()
+    print("Patching.... Pls wait") 
      
     with open(icon_path, 'rb') as fp:
         png_content = fp.read()
@@ -148,57 +185,74 @@ def changeIcon():
 
     with open(plist_path, 'wb') as fp:
         plistlib.dump(plist, fp)
+        
+    zip_ipa(ipa_path, app_path, file_name_no_ipa)
+    clear_terminal()
+    print("The App Icon was changed successfully.")
     
+if option == 5:
+
+    try:
+        with open(filename, 'rb') as f:
+            satella_jailed_folder = pickle.load(f)
+    except FileNotFoundError:
+        satella_jailed_folder = input("enter the folder where Satella Jailed was downloaded: \n")
+        clear_terminal()
+        with open(filename, 'wb') as f:
+            pickle.dump(satella_jailed_folder, f)
+        print('created permanent variable for your Satella path.\n')
+        time.sleep(4)
+        clear_terminal()
+
+    ipa_path = input("enter the path of the .iPA you want to inject Satella to: \n")
+    clear_terminal()
+    shutil.copy(ipa_path, satella_jailed_folder)
+    os.chdir(satella_jailed_folder)
+    patchsh_path = satella_jailed_folder + "/patch.sh"
+    print("Patching.... Pls wait a few seconds\n")
+    #subprocess.run("chmod +x " + patchsh_path)
+    process = subprocess.Popen(['sh', 'patch.sh'], stdout=subprocess.PIPE, text=True)
+    while True:
+        output = process.stdout.readline()
+        if output == '' and process.poll() is not None:
+            break
+        if output:
+            print(output.strip())
+    process.communicate()
+    clear_terminal()
+    print(".iPA file with Satella injected should be here: " + satella_jailed_folder)
     
-    payload_path2 = payload_path
-    
-    if os.path.basename(payload_path2) == "Payload":
-        output_path = payload_path2[:len(payload_path2)-len("/Payload")]
+if option == 6:
+        
+    azule_ipa_input = input("enter the path to the iPA you want to inject debs into: \n")
+    clear_terminal()
+    if os.path.isfile(azule_ipa_input):
+        print(".iPA file exists.\n\n")
+        time.sleep(4)
+        clear_terminal()
     else:
-        output_path = os.path.dirname(payload_path2)
- 
-    output_path = os.path.join(output_path) 
-    with zipfile.ZipFile(os.path.join(output_path, "Payload.zip"), 'w', zipfile.ZIP_DEFLATED) as zip_file:  
-       for root, dirs, files in os.walk(payload_path): 
-           for file in files: 
-               file_path = os.path.join(root, file) 
-               zip_file.write(file_path, file_path.replace(payload_path, "Payload")) 
-                
-                
-    user_new_ipa_name = input(f"\n\nenter a new name for your edited .ipa (without the .ipa at the end) \noriginal .ipa name: {file_name_no_ipa}'\n") 
-    user_new_ipa_name = user_new_ipa_name.strip() + ".ipa"
-    os.rename(payload_path+".zip", user_new_ipa_name) 
-    edited_file_path = os.path.join(output_path, user_new_ipa_name)
-    os.replace(user_new_ipa_name, edited_file_path)
-     
-    print("The App-Icon was changed successfully.") 
- 
-ipa_path = input("enter the path to your .ipa: ") 
- 
-file_name = os.path.basename(ipa_path) 
-file_name_no_ipa = file_name[:-4]
- 
-zip_path = ipa_path.replace(".ipa", ".zip") 
- 
-if os.path.exists(ipa_path): 
-    os.rename(ipa_path, zip_path) 
-    print(".ipa file successfully renamed to .zip.") 
-else: 
-    print("The .ipa file could not be found. Try again...") 
-    exit() 
- 
-print("[1] change Bundle-ID") 
-print("[2] change App-Name") 
-print("[3] change App-Icon") 
- 
-option = int(input("choose an option: ")) 
- 
-if option == 1: 
-    bundleID() 
- 
-if option == 2: 
-     changeName() 
- 
-if option == 3:
-    changeIcon()
+        print("couldnt find .iPA file. Try again.")
+        exit()
+        
+    azule_ipa_output = input("enter the output path for your modified .iPA: \n")
+    clear_terminal()
+    azule_ipa_output_name = input("enter new name for the modified .iPA (without the .iPA at the end)\n")
+    clear_terminal()
+    file_paths = []
+    while True:
+        path = input("enter path of the deb you want to inject (type 'ready' if you are done) \n")
+        if path == "done":
+            break
+        clear_terminal()
+        file_paths.append(path)
+
+    deb_paths = " ".join(file_paths)
+    terminal_command = f"azule -o '{azule_ipa_output}' -i '{azule_ipa_input}' -f '{deb_paths}' -z -n '{azule_ipa_output_name}'"
+
+    subprocess.run(terminal_command, shell=True)
+    clear_terminal()
+    print("modified .iPA should be here:" + azule_ipa_output)
     
+if option > 6:
+    print("not a valid option. Try again.")
+    exit()
